+++ common/datavalues/src/arrays/null/iterator.rs
@@ -1,0 +1,43 @@
+// Copyright 2020 Datafuse Labs.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+
+
+/// Trait for DFArrays that don't have null values.
+/// The result is the most efficient implementation `Iterator`, according to the number of chunks.
+pub trait IntoNoNullIterator {
+    type Item;
+    type IntoIter: Iterator<Item = Self::Item>;
+
+    fn into_no_null_iter(self) -> Self::IntoIter;
+}
+
+/// Wrapper struct to convert an iterator of type `T` into one of type `Option<T>`.  It is useful to make the
+/// `IntoIterator` trait, in which every iterator shall return an `Option<T>`.
+pub struct SomeIterator<I>(I)
+where I: Iterator;
+
+impl<I> Iterator for SomeIterator<I>
+where I: Iterator
+{
+    type Item = Option<I::Item>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        self.0.next().map(Some)
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        self.0.size_hint()
+    }
+}
