+++ common/datavalues/src/arrays/null/mod.rs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod iterator;
 mod null;
 
+pub use iterator::*;
 pub use null::*;
