+++ common/datavalues/src/arrays/primitive/primitive.rs
@@ -15,6 +15,7 @@
 use common_arrow::arrow::array::Array;
 use common_arrow::arrow::array::PrimitiveArray;
 use common_arrow::arrow::bitmap::Bitmap;
+use common_arrow::arrow::buffer::Buffer;
 use common_exception::ErrorCode;
 use common_exception::Result;
 
@@ -178,3 +179,84 @@ pub type DFInt64Array = DFPrimitiveArray<i64>;
 
 pub type DFFloat32Array = DFPrimitiveArray<f32>;
 pub type DFFloat64Array = DFPrimitiveArray<f64>;
+
+/// # Safety
+/// Note this doesn't do any bound checking, for performance reason.
+/// Take kernel for single chunk without nulls and an iterator as index.
+pub unsafe fn take_primitive_iter_unchecked<T: DFPrimitiveType, I: IntoIterator<Item = usize>>(
+    arr: &PrimitiveArray<T>,
+    indices: I,
+) -> Arc<PrimitiveArray<T>> {
+    match arr.null_count() {
+        0 => {
+            let array_values = arr.values().as_slice();
+            let iter = indices
+                .into_iter()
+                .map(|idx| *array_values.get_unchecked(idx));
+
+            let values = Buffer::from_trusted_len_iter_unchecked(iter);
+            Arc::new(PrimitiveArray::from_data(
+                T::data_type().to_arrow(),
+                values,
+                None,
+            ))
+        }
+        _ => {
+            let array_values = arr.values();
+
+            let arr = indices
+                .into_iter()
+                .map(|idx| {
+                    if arr.is_valid(idx) {
+                        Some(array_values[idx])
+                    } else {
+                        None
+                    }
+                })
+                .collect();
+
+            Arc::new(arr)
+        }
+    }
+}
+
+/// # Safety
+/// Note this doesn't do any bound checking, for performance reason.
+/// Take kernel for a single chunk and an iterator that can produce None values.
+/// This is used in join operations.
+pub unsafe fn take_primitive_opt_iter_unchecked<
+    T: DFPrimitiveType,
+    I: IntoIterator<Item = Option<usize>>,
+>(
+    arr: &PrimitiveArray<T>,
+    indices: I,
+) -> Arc<PrimitiveArray<T>> {
+    match arr.null_count() {
+        0 => {
+            let array_values = arr.values();
+
+            let iter = indices
+                .into_iter()
+                .map(|opt_idx| opt_idx.map(|idx| *array_values.get_unchecked(idx)));
+            let arr = PrimitiveArray::from_trusted_len_iter_unchecked(iter);
+
+            Arc::new(arr)
+        }
+        _ => {
+            let array_values = arr.values();
+
+            let iter = indices.into_iter().map(|opt_idx| {
+                opt_idx.and_then(|idx| {
+                    if arr.is_valid(idx) {
+                        Some(*array_values.get_unchecked(idx))
+                    } else {
+                        None
+                    }
+                })
+            });
+            let arr = PrimitiveArray::from_trusted_len_iter_unchecked(iter);
+
+            Arc::new(arr)
+        }
+    }
+}
