+++ common/datavalues/src/arrays/utf8/builder_test.rs
@@ -19,7 +19,7 @@ fn test_empty_array() {
     let mut builder = Utf8ArrayBuilder::with_capacity(16);
     let data_array = builder.finish();
     assert_eq!(true, data_array.is_empty());
-    assert_eq!(DataType::Utf8, data_array.data_type());
+    assert_eq!(&DataType::Utf8, data_array.data_type());
 }
 
 #[test]
@@ -29,7 +29,7 @@ fn test_fill_data() {
     builder.append_option(Some("\u{1F378}"));
     builder.append_null();
 
-    let data_array: DataArray<Utf8Type> = builder.finish();
+    let data_array = builder.finish();
     let mut iter = data_array.into_iter();
 
     assert_eq!(3, data_array.len());
@@ -41,7 +41,7 @@ fn test_fill_data() {
 
 #[test]
 fn test_new_from_opt_slice() {
-    let data_array: DataArray<Utf8Type> = DataArray::new_from_opt_slice(&[Some("你好"), None]);
+    let data_array = DFUtf8Array::new_from_opt_slice(&[Some("你好"), None]);
     let mut iter = data_array.into_iter();
 
     assert_eq!(2, data_array.len());
@@ -55,7 +55,7 @@ fn test_new_from_opt_iter() {
     let v = vec![None, Some("你好"), None];
     let mut iter = v.into_iter();
     iter.next(); // move iterator and create data array from second element
-    let data_array: DataArray<Utf8Type> = DataArray::new_from_opt_iter(iter);
+    let data_array = DFUtf8Array::new_from_opt_iter(iter);
     let mut iter = data_array.into_iter();
 
     assert_eq!(2, data_array.len());
